feat: add preliminary speech recognition
On the first turn of a 2 player game, the game will listen for a word
and correctly place an X on the center when it detects the word
'center' being said. Only the first turn though, still a lot more work
to do and it's too slow as well.

Renamed whisper_model.py to speech_rec.py and initialized the
recognizer and microphone there as well, instead of inside the TTTGame
class.