feat: listen for speech on every turn and disable mouse
The program now listens for speech every turn (for the 2 player) option,
so it (theoretically) allows for a game to be played between two
players only using speech.

It turns out it is hard to accept both speech input and mouse input and
prevent the two from messing with each other, so an audio_only flag was
created and set to True for the time being. Allowing the user to use
either one is a future goal.

The model for whisper was increased to 'base' which I hoped would lead
to better accuracy, but that is still to be determined.

Lastly, the CELL_HOTWORDS dictionary has been expanded so the user can
use a couple more words to indicate a cell.